% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variability.R
\name{check_variability}
\alias{check_variability}
\title{Check Variability}
\usage{
check_variability(data = NULL)
}
\arguments{
\item{data}{raw data files}
}
\value{
Returns variability check.
}
\description{
Check Variability
}
\keyword{internal}
