% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_coef.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Extract Beta Coefficients.}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ridgereg}}

\item{...}{other arguments.}
}
\value{
Coefficients extracted from the model \code{\link{ridgereg}} and returned as a named vector.
}
\description{
\code{coef} returns the estimated beta coefficients of the regression model calculated by the \code{\link{ridgereg}} function.
}
\examples{
x<- ridgereg(formula= Sepal.Width ~ Sepal.Length, data= iris)
coef(x)
}

