% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactions.R
\name{create_interactions}
\alias{create_interactions}
\title{Create Interactions}
\usage{
create_interactions(endog = NULL, exog = NULL, interactions = NULL)
}
\arguments{
\item{endog}{Matrix of endogenous variables (at time and lagged).}

\item{exog}{Matrix of exogenous variables (at time and lagged).}

\item{interactions}{List of specified interactions.}
}
\value{
Returns subject data with new interactions and centered main effects if desired.
}
\description{
Create Interactions
}
\keyword{internal}
