% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiREG-data.R
\docType{data}
\name{examplesim}
\alias{examplesim}
\title{Toy example for documentation}
\format{A list of data frames with 3 individuals, who each have 500 observations on 5 variables.}
\usage{
examplesim
}
\description{
This object contains a list of time series data for 3 individuals 
with 500 time points and 5 variables.
}
\keyword{datasets}
