% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_search.R
\name{group_search}
\alias{group_search}
\title{Find subgroup solutions using final patterns.}
\usage{
group_search(subdata, groupcutoff, yvarnames, interactnames, output,
  grppen = NULL, initial_penalties = NULL, verbose)
}
\arguments{
\item{subdata}{Data from previous step.}

\item{groupcutoff}{Thresholds for group-level paths.}

\item{yvarnames}{Endogenous variable names.}

\item{interactnames}{Names of interaction variables.}

\item{output}{Output object}

\item{grppen}{Group path penalties.}

\item{initial_penalties}{User-set penalties.}

\item{verbose}{Logical. If TRUE, algorithm will print progress to console.}
}
\value{
Returns subgroup membership, similarity matrix, modularity, and number of subgroups.
}
\description{
Find subgroup solutions using final patterns.
}
\keyword{internal}
