% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderated_group_paths.R
\name{group_subgroupsign}
\alias{group_subgroupsign}
\title{Identify Moderated Group Paths}
\usage{
group_subgroupsign(group_paths, finalpaths, subgroup_mem)
}
\arguments{
\item{group_paths}{Matrix indicating which paths are group-level.}

\item{finalpaths}{Matrix of final coefficients for each path.}

\item{subgroup_mem}{Vector of subgroup membership.}
}
\value{
Returns indices of moderated group paths.
}
\description{
Identify Moderated Group Paths
}
\keyword{internal}
