% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_search.R
\name{ind_search}
\alias{ind_search}
\title{Find individual-level solutions using final patterns.}
\usage{
ind_search(subdata, yvarnames, interactnames, grppen, output, verbose)
}
\arguments{
\item{subdata}{Data from previous step.}

\item{yvarnames}{Endogenous variable names.}

\item{interactnames}{Names of interaction variables.}

\item{grppen}{Group path penalties.}

\item{output}{Output object}

\item{verbose}{Logical. If TRUE, algorithm will print progress to console.}
}
\value{
Returns final paths for all individuals.
}
\description{
Find individual-level solutions using final patterns.
}
\keyword{internal}
