% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2state.R
\name{latlong2state}
\alias{latlong2state}
\title{Converting latitude and longitude to state names}
\usage{
latlong2state(pointsDF)
}
\arguments{
\item{pointsDF}{A data frame which column 1 contains the longitude in degrees (negative in US). 
column 2 contains the latitude.}
}
\value{
a character vector with the US states names.
}
\description{
The function converts latitude and longitude to US state names. Especially used in function \code{visualize_airport_delay}
}
\examples{
testPoints <- data.frame(x = c(-90, -120), y = c(44, 44))
latlong2state(testPoints)
}

