% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Model Selection for Regularized Parameters}
\usage{
model_selection(x = NULL, y = NULL, selection_crit = NULL,
  alpha = NULL, penalty.factor = NULL)
}
\description{
Model Selection for Regularized Parameters
}
\keyword{internal}
