\name{multireg-package}
\alias{multireg-package}
\alias{multireg}
\docType{package}
\title{
\packageTitle{multireg}
}
\description{
\packageDescription{multireg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{multireg}
\packageIndices{multireg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{multireg}

Maintainer: \packageMaintainer{multireg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
