% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot residuals.}
\usage{
plot.linreg(x, ...)
}
\arguments{
\item{x}{An object of class linreg.}

\item{...}{other arguments.}
}
\value{
A two in one plot over residuals versus fitted values.
}
\description{
The function plots the residuals versus fitted values, and the standardized residuals
versus the fitted value
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
plot(x)
}

