% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot residuals.}
\usage{
plot(x)
}
\arguments{
\item{argument}{An object of class linreg.}
}
\value{
A two in one plot over residuals versus fitted values.
}
\description{
The function plots the residuals versus fitted values, and the standardized residuals
versus the fitted value
}
\examples{
plot(x=linreg)
}

