% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Fitted values of the regression model.}
\usage{
pred.linreg(x)
}
\arguments{
\item{See}{\code{\link{linreg}}}
}
\value{
A named vector.
}
\description{
Returning the fitted values of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
pred(x)
}

