% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Extract Fitted Values from linreg object.}
\usage{
\method{predict}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}}

\item{...}{other arguments.}
}
\value{
A numeric vector of fitted values extracted from the model in \code{\link{linreg}}.
}
\description{
\code{pred} returns the fitted values of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
predict(x)
}

