% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Extract Beta Coefficients and equation if the model.}
\usage{
\method{print}{linreg}(x)
}
\arguments{
\item{See}{\code{\link{linreg}}}
}
\value{
Coefficients and equations extracted from the model \code{x} and returned as a named vector.
}
\description{
\code{coef} returns the estimated beta coefficients and the equation of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
print(x)
}

