% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_print.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Extract Beta Coefficients and equation if the model.}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ridgereg}}

\item{...}{other arguments.}
}
\value{
Coefficients and equations extracted from the model in \code{\link{ridgereg}} and returned as a named vector.
}
\description{
\code{print} returns the estimated beta coefficients and the equation of the regression model calculated by the \code{\link{ridgereg}} function.
}
\examples{
x<- ridgereg(formula= Sepal.Width ~ Sepal.Length, data= iris)
print(x)
}

