% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Extract Residuals from linreg object.}
\usage{
\method{resid}{linreg}(x)
}
\arguments{
\item{x}{is an object of the class \code{linreg}.}
}
\value{
A numeric vector with the residuals extracted from the model calculated by \code{\link{linreg}} function.
}
\description{
\code{resid} returns the residuals of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
resid(x)
}

