% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_search.R
\name{subgroup_search}
\alias{subgroup_search}
\title{Find subgroup solutions using final patterns.}
\usage{
subgroup_search(subdata, indpaths, output, verbose)
}
\arguments{
\item{subdata}{Data from previous step.}

\item{indpaths}{Results from individual-level path search.}

\item{output}{Output object.}

\item{verbose}{Logical. If TRUE, algorithm will print progress to console.}
}
\value{
Returns subgroup membership, similarity matrix, modularity, and number of subgroups.
}
\description{
Find subgroup solutions using final patterns.
}
\keyword{internal}
