% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Extract Summary from linreg object.}
\usage{
\method{summary}{linreg}(x)
}
\arguments{
\item{See}{\code{\link{linreg}}}
}
\value{
Summary extracted from the model \code{x}.
}
\description{
\code{summary} returns the fitted values of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
summary(x)
}

