% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airports_delay}
\alias{visualize_airports_delay}
\title{Visualizing using a choropleth map the mean total flight delays.}
\usage{
visualize_airports_delay()
}
\value{
A ggplot object visualizes into a choropleth map.
}
\description{
The function visualizes the mean total flight delays from different aiports in the US. By using the data \code{airports} and \code{flights} from the package \code{nycflights13}, the mean total time delays were calculated.
}

