% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract Beta Coefficients.}
\usage{
\method{coef}{linreg}(x)
}
\arguments{
\item{See}{\code{\link{linreg}}}
}
\value{
Coefficients extracted from the model \code{x} and returned as a named vector.
}
\description{
\code{coef} returns the estimated beta coefficients of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
coef(x)
}

