% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\title{Beta coefficients of the regression model.}
\usage{
coef(x)
}
\arguments{
\item{See}{\code{\link{linreg}}}
}
\value{
A named vector.
}
\description{
Returning the beta estimates of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
coef(x)
}

