% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_search.R
\name{group_search}
\alias{group_search}
\title{Find subgroup solutions using final patterns.}
\usage{
group_search(subdata, groupcutoff, yvarnames, interact_exogenous,
  predict_with_interactions, interactnames, interact_exogvars, output,
  grppen = NULL, initial_penalties = NULL)
}
\arguments{
\item{subdata}{Data from previous step.}

\item{groupcutoff}{Thresholds for group-level paths.}

\item{yvarnames}{Endogenous variable names.}

\item{interact_exogenous}{Exogenous variable to use as moderators.}

\item{predict_with_interactions}{Endogenous variables to be predicted by interactions.}

\item{interactnames}{Names of interaction variables.}

\item{interact_exogvars}{Exogenous variable to use as moderators.}

\item{output}{Output object}

\item{grppen}{Group path penalties.}

\item{initial_penalties}{User-set penalties.}
}
\value{
Returns subgroup membership, similarity matrix, modularity, and number of subgroups.
}
\description{
Find subgroup solutions using final patterns.
}
\keyword{internal}
