% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_search.R
\name{ind_search}
\alias{ind_search}
\title{Find individual-level solutions using final patterns.}
\usage{
ind_search(subdata, yvarnames, interact_exogenous,
  predict_with_interactions, interactnames, interact_exogvars, grppen,
  output)
}
\value{
Returns final paths for all individuals.
}
\description{
Find individual-level solutions using final patterns.
}
\keyword{internal}
