% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS.R
\name{linreg}
\alias{linreg}
\title{Computing regression statistics using ordinary least squares.}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A regression formula.}

\item{data}{A dataframe containing the variables for used in formula.}
}
\value{
An object of class linreg.
}
\description{
The function calculates the regression coefficients, fitted values, residuals, 
degrees of freedom, residual variance, 
variance of regression coefficients, t-values and p-values for each coefficients.
}
\examples{
linreg(forumla= Sepal.Width ~ Sepal.Length, data= iris)
linreg(formula = eruptions ~ waiting, data= faithful)
}

