% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_output.R
\name{manage_output}
\alias{manage_output}
\title{Manage Output from multiREG}
\usage{
manage_output(out = NULL, plot = NULL, output = NULL,
  verbose = TRUE)
}
\arguments{
\item{out}{Controls whether output should be written to file.}

\item{plot}{Controls whether plots should be generated.}

\item{output}{Output object.}

\item{verbose}{Logical. If TRUE, algorithm will print progress to console.}
}
\value{
Returns output object and prints output files if necessary.
}
\description{
Manage Output from multiREG
}
\keyword{internal}
