% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multireg-package.R
\docType{package}
\name{multireg-package}
\alias{multireg}
\alias{multireg-package}
\title{multireg: A module for demonstrating multiple regression and ANCOVA}
\description{
This module is used to demonstrate building a package that contains a vignette about running multiple regression using an included dataset.
}
\author{
\strong{Maintainer}: Anthony Di Fiore \email{anthony.difiore@austin.utexas.edu} (YOUR-ORCID-ID)

}
\keyword{internal}
