% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{network_visualization}
\alias{network_visualization}
\title{Network Visualization}
\usage{
network_visualization(output = NULL)
}
\arguments{
\item{output}{Output object.}
}
\value{
Returns network plots.
}
\description{
Network Visualization
}
\keyword{internal}
