% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot residuals.}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg.}
}
\value{
A two in one plot over residuals versus fitted values.
}
\description{
The function plots the residuals versus fitted values, and the standardized residuals
versus the fitted value
}
\examples{
plot(x=linreg)
}

