% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Extract Fitted Values from linreg object.}
\usage{
\method{predict}{linreg}(x)
}
\arguments{
\item{See}{\code{\link{linreg}}}
}
\value{
Fitted values extracted from the model \code{x}.
}
\description{
\code{pred} returns the fitted values of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
pred(x)
}

