% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Extract Beta Coefficients and equation if the model.}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{linreg}}

\item{...}{other arguments.}
}
\value{
Coefficients and equations extracted from the model in \code{\link{linreg}} and returned as a named vector.
}
\description{
\code{print} returns the estimated beta coefficients and the equation of the regression model calculated by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
print(x)
}

