% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Computing bridge regression using ordinary least squares.}
\usage{
ridgereg(formula, data, lambda = 0)
}
\arguments{
\item{formula}{A regression formula.}

\item{data}{A dataframe containing the variables used in formula.}

\item{Optional}{choiche of lambda, default = 0.}
}
\value{
An object of class ridgereg.
}
\description{
The function calculates the regression coefficients, and fitted values
}
\examples{
ridgereg(formula= Sepal.Width ~ Sepal.Length + Petal.Width, data= iris)
}

