% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Extract Summary Results from linreg object.}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linreg}}

\item{...}{other arguments.}
}
\value{
A list of summary result extracted from the model in \code{\link{linreg}}.
}
\description{
\code{summary} returns the result summariest of the regression model fitted by the \code{\link{linreg}} function.
}
\examples{
x<- linreg(formula= Sepal.Width ~ Sepal.Length, data= iris)
summary(x)
}

